skopt_requirement.txt
absl-py                 0.15.0
alembic                 1.8.0
antlr4-python3-runtime  4.9.3
asttokens               2.0.5
astunparse              1.6.3
attrs                   21.4.0
autopage                0.5.1
backcall                0.2.0
bayesian-optimization   1.2.0
cachetools              5.1.0
certifi                 2022.5.18.1
charset-normalizer      2.0.12
cliff                   3.10.1
cmaes                   0.8.2
cmd2                    2.4.1
colorlog                6.6.0
cycler                  0.11.0
decorator               5.1.1
efficientnet            1.1.0
executing               0.8.3
flatbuffers             1.12
fonttools               4.33.3
gast                    0.3.3
google-auth             2.6.6
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
greenlet                1.1.2
grpcio                  1.32.0
h5py                    2.10.0
hydra-core              1.2.0
idna                    3.3
imageio                 2.19.2
importlib-metadata      4.11.4
importlib-resources     5.7.1
ipython                 8.3.0
jedi                    0.18.1
joblib                  1.1.0
Keras                   2.4.3
Keras-Applications      1.0.8
keras-efficientnet      0.1.4
Keras-Preprocessing     1.1.2
keras-tuner             1.1.2
kiwisolver              1.4.2
kt-legacy               1.0.4
Mako                    1.2.0
Markdown                3.3.7
MarkupSafe              2.1.1
matplotlib              3.5.2
matplotlib-inline       0.1.3
networkx                2.8.2
numpy                   1.19.5
oauthlib                3.2.0
omegaconf               2.2.1
opencv-python           4.6.0.66
opt-einsum              3.3.0
optuna                  2.10.0
packaging               21.3
pandas                  1.4.2
parso                   0.8.3
pbr                     5.9.0
pickleshare             0.7.5
Pillow                  9.1.1
pip                     21.2.4
prettytable             3.3.0
prompt-toolkit          3.0.29
protobuf                3.20.1
pure-eval               0.2.2
pyaml                   21.10.1
pyasn1                  0.4.8
pyasn1-modules          0.2.8
Pygments                2.12.0
pyparsing               3.0.9
pyperclip               1.8.2
python-dateutil         2.8.2
pytz                    2022.1
PyWavelets              1.3.0
PyYAML                  6.0
requests                2.28.0
requests-oauthlib       1.3.1
rsa                     4.8
scikit-image            0.19.2
scikit-learn            1.1.1
scikit-optimize         0.9.0
scipy                   1.8.1
setuptools              61.2.0
six                     1.15.0
SQLAlchemy              1.4.37
stack-data              0.2.0
stevedore               3.5.0
tensorboard             2.9.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorflow              2.4.1
tensorflow-addons       0.17.0
tensorflow-estimator    2.4.0
termcolor               1.1.0
threadpoolctl           3.1.0
tifffile                2022.5.4
tqdm                    4.64.0
traitlets               5.2.1.post0
typeguard               2.13.3
typing-extensions       3.7.4.3
urllib3                 1.26.9
wcwidth                 0.2.5
Werkzeug                2.1.2
wheel                   0.37.1
wrapt                   1.12.1
zipp                    3.8.0
